+++ exporters/prometheus/src/exporter_utils.cc
@@ -1,14 +1,22 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+#include <limits>
 #include <sstream>
+#include <string>
+#include <unordered_map>
+#include <unordered_set>
 #include <utility>
 #include <vector>
-#include "prometheus/metric_family.h"
 
-#include <prometheus/metric_type.h>
+#include "prometheus/metric_family.h"
+#include "prometheus/metric_type.h"
+
 #include "opentelemetry/exporters/prometheus/exporter_utils.h"
 #include "opentelemetry/sdk/metrics/export/metric_producer.h"
+#include "opentelemetry/sdk/resource/resource.h"
+#include "opentelemetry/sdk/resource/semantic_conventions.h"
+#include "opentelemetry/trace/semantic_conventions.h"
 
 #include "opentelemetry/sdk/common/global_log_handler.h"
 
@@ -22,6 +30,10 @@ namespace metrics
 {
 namespace
 {
+
+static constexpr const char *kScopeNameKey    = "otel_scope_name";
+static constexpr const char *kScopeVersionKey = "otel_scope_version";
+
 /**
  * Sanitize the given metric name by replacing invalid characters with _,
  * ensuring that multiple consecutive _ characters are collapsed to a single _.
@@ -107,11 +119,25 @@ std::string SanitizeName(std::string name)
  * @return a collection of translated metrics that is acceptable by Prometheus
  */
 std::vector<prometheus_client::MetricFamily> PrometheusExporterUtils::TranslateToPrometheus(
-    const sdk::metrics::ResourceMetrics &data)
+    const sdk::metrics::ResourceMetrics &data,
+    bool populate_target_info)
 {
 
   // initialize output vector
+  std::size_t reserve_size = 1;
+  for (const auto &instrumentation_info : data.scope_metric_data_)
+  {
+    reserve_size += instrumentation_info.metric_data_.size();
+  }
+
   std::vector<prometheus_client::MetricFamily> output;
+  output.reserve(reserve_size);
+
+  // Append target_info as the first metric
+  if (populate_target_info && !data.scope_metric_data_.empty())
+  {
+    SetTarget(data, (*data.scope_metric_data_.begin()).scope_, &output);
+  }
 
   for (const auto &instrumentation_info : data.scope_metric_data_)
   {
@@ -147,10 +173,11 @@ std::vector<prometheus_client::MetricFamily> PrometheusExporterUtils::TranslateT
           }
           else
           {
-            sum = nostd::get<int64_t>(histogram_point_data.sum_);
+            sum = static_cast<double>(nostd::get<int64_t>(histogram_point_data.sum_));
           }
           SetData(std::vector<double>{sum, (double)histogram_point_data.count_}, boundaries, counts,
-                  point_data_attr.attributes, &metric_family);
+                  point_data_attr.attributes, instrumentation_info.scope_, &metric_family,
+                  data.resource_);
         }
         else if (type == prometheus_client::MetricType::Gauge)
         {
@@ -160,14 +187,16 @@ std::vector<prometheus_client::MetricFamily> PrometheusExporterUtils::TranslateT
             auto last_value_point_data =
                 nostd::get<sdk::metrics::LastValuePointData>(point_data_attr.point_data);
             std::vector<metric_sdk::ValueType> values{last_value_point_data.value_};
-            SetData(values, point_data_attr.attributes, type, &metric_family);
+            SetData(values, point_data_attr.attributes, instrumentation_info.scope_, type,
+                    &metric_family, data.resource_);
           }
           else if (nostd::holds_alternative<sdk::metrics::SumPointData>(point_data_attr.point_data))
           {
             auto sum_point_data =
                 nostd::get<sdk::metrics::SumPointData>(point_data_attr.point_data);
             std::vector<metric_sdk::ValueType> values{sum_point_data.value_};
-            SetData(values, point_data_attr.attributes, type, &metric_family);
+            SetData(values, point_data_attr.attributes, instrumentation_info.scope_, type,
+                    &metric_family, data.resource_);
           }
           else
           {
@@ -183,7 +212,8 @@ std::vector<prometheus_client::MetricFamily> PrometheusExporterUtils::TranslateT
             auto sum_point_data =
                 nostd::get<sdk::metrics::SumPointData>(point_data_attr.point_data);
             std::vector<metric_sdk::ValueType> values{sum_point_data.value_};
-            SetData(values, point_data_attr.attributes, type, &metric_family);
+            SetData(values, point_data_attr.attributes, instrumentation_info.scope_, type,
+                    &metric_family, data.resource_);
           }
           else
           {
@@ -197,6 +227,17 @@ std::vector<prometheus_client::MetricFamily> PrometheusExporterUtils::TranslateT
     }
   }
   return output;
+}
+
+void PrometheusExporterUtils::AddPrometheusLabel(
+    std::string name,
+    std::string value,
+    std::vector<::prometheus::ClientMetric::Label> *labels)
+{
+  prometheus_client::ClientMetric::Label prometheus_label;
+  prometheus_label.name  = std::move(name);
+  prometheus_label.value = std::move(value);
+  labels->emplace_back(std::move(prometheus_label));
 }
 
 metric_sdk::AggregationType PrometheusExporterUtils::getAggregationType(
@@ -252,19 +293,53 @@ prometheus_client::MetricType PrometheusExporterUtils::TranslateType(
   }
 }
 
+void PrometheusExporterUtils::SetTarget(
+    const sdk::metrics::ResourceMetrics &data,
+    const opentelemetry::sdk::instrumentationscope::InstrumentationScope *scope,
+    std::vector<::prometheus::MetricFamily> *output)
+{
+  if (output == nullptr || data.resource_ == nullptr)
+  {
+    return;
+  }
+
+  prometheus_client::MetricFamily metric_family;
+  metric_family.name = "target";
+  metric_family.help = "Target metadata";
+  metric_family.type = prometheus_client::MetricType::Info;
+  metric_family.metric.emplace_back();
+
+  prometheus_client::ClientMetric &metric = metric_family.metric.back();
+  metric.info.value                       = 1.0;
+
+  metric_sdk::PointAttributes empty_attributes;
+  SetMetricBasic(metric, empty_attributes, scope, data.resource_);
+
+  for (auto &label : data.resource_->GetAttributes())
+  {
+    AddPrometheusLabel(SanitizeName(label.first), AttributeValueToString(label.second),
+                       &metric.label);
+  }
+
+  output->emplace_back(std::move(metric_family));
+}
+
 /**
  * Set metric data for:
  * sum => Prometheus Counter
  */
 template <typename T>
-void PrometheusExporterUtils::SetData(std::vector<T> values,
-                                      const metric_sdk::PointAttributes &labels,
-                                      prometheus_client::MetricType type,
-                                      prometheus_client::MetricFamily *metric_family)
+void PrometheusExporterUtils::SetData(
+    std::vector<T> values,
+    const metric_sdk::PointAttributes &labels,
+    const opentelemetry::sdk::instrumentationscope::InstrumentationScope *scope,
+    prometheus_client::MetricType type,
+    prometheus_client::MetricFamily *metric_family,
+    const opentelemetry::sdk::resource::Resource *resource)
 {
   metric_family->metric.emplace_back();
   prometheus_client::ClientMetric &metric = metric_family->metric.back();
-  SetMetricBasic(metric, labels);
+  SetMetricBasic(metric, labels, scope, resource);
   SetValue(values, type, &metric);
 }
 
@@ -273,25 +348,31 @@ void PrometheusExporterUtils::SetData(std::vector<T> values,
  * Histogram => Prometheus Histogram
  */
 template <typename T>
-void PrometheusExporterUtils::SetData(std::vector<T> values,
-                                      const std::vector<double> &boundaries,
-                                      const std::vector<uint64_t> &counts,
-                                      const metric_sdk::PointAttributes &labels,
-                                      prometheus_client::MetricFamily *metric_family)
+void PrometheusExporterUtils::SetData(
+    std::vector<T> values,
+    const std::vector<double> &boundaries,
+    const std::vector<uint64_t> &counts,
+    const metric_sdk::PointAttributes &labels,
+    const opentelemetry::sdk::instrumentationscope::InstrumentationScope *scope,
+    prometheus_client::MetricFamily *metric_family,
+    const opentelemetry::sdk::resource::Resource *resource)
 {
   metric_family->metric.emplace_back();
   prometheus_client::ClientMetric &metric = metric_family->metric.back();
-  SetMetricBasic(metric, labels);
+  SetMetricBasic(metric, labels, scope, resource);
   SetValue(values, boundaries, counts, &metric);
 }
 
 /**
  * Set labels to metric data
  */
-void PrometheusExporterUtils::SetMetricBasic(prometheus_client::ClientMetric &metric,
-                                             const metric_sdk::PointAttributes &labels)
+void PrometheusExporterUtils::SetMetricBasic(
+    prometheus_client::ClientMetric &metric,
+    const metric_sdk::PointAttributes &labels,
+    const opentelemetry::sdk::instrumentationscope::InstrumentationScope *scope,
+    const opentelemetry::sdk::resource::Resource *resource)
 {
-  if (labels.empty())
+  if (labels.empty() && nullptr == resource)
   {
     return;
   }
@@ -300,7 +381,7 @@ void PrometheusExporterUtils::SetMetricBasic(prometheus_client::ClientMetric &me
   // Note that attribute keys are sorted, but sanitized keys can be out-of-order.
   // We could sort the sanitized keys again, but this seems too expensive to do
   // in this hot code path. Instead, we ignore out-of-order keys and emit a warning.
-  metric.label.reserve(labels.size());
+  metric.label.reserve(labels.size() + 2);
   std::string previous_key;
   for (auto const &label : labels)
   {
@@ -323,6 +404,24 @@ void PrometheusExporterUtils::SetMetricBasic(prometheus_client::ClientMetric &me
           << label.first << "' became '" << sanitized << "' which is less than '" << previous_key
           << "'. Ignoring this label.");
     }
+  }
+  if (!scope)
+  {
+    return;
+  }
+  auto scope_name = scope->GetName();
+  if (!scope_name.empty())
+  {
+    metric.label.emplace_back();
+    metric.label.back().name  = kScopeNameKey;
+    metric.label.back().value = std::move(scope_name);
+  }
+  auto scope_version = scope->GetVersion();
+  if (!scope_version.empty())
+  {
+    metric.label.emplace_back();
+    metric.label.back().name  = kScopeVersionKey;
+    metric.label.back().value = std::move(scope_version);
   }
 }
 
@@ -379,7 +478,7 @@ void PrometheusExporterUtils::SetValue(std::vector<T> values,
   const auto &value_var = values[0];
   if (nostd::holds_alternative<int64_t>(value_var))
   {
-    value = nostd::get<int64_t>(value_var);
+    value = static_cast<double>(nostd::get<int64_t>(value_var));
   }
   else
   {
@@ -414,9 +513,9 @@ void PrometheusExporterUtils::SetValue(std::vector<T> values,
                                        const std::vector<uint64_t> &counts,
                                        prometheus_client::ClientMetric *metric)
 {
-  metric->histogram.sample_sum   = values[0];
-  metric->histogram.sample_count = values[1];
-  int cumulative                 = 0;
+  metric->histogram.sample_sum   = static_cast<double>(values[0]);
+  metric->histogram.sample_count = static_cast<std::uint64_t>(values[1]);
+  std::uint64_t cumulative       = 0;
   std::vector<prometheus_client::ClientMetric::Bucket> buckets;
   uint32_t idx = 0;
   for (const auto &boundary : boundaries)