+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Increment the:
 
 * [DEPRECATION] Deprecate ZPAGES
   [#2291](https://github.com/open-telemetry/opentelemetry-cpp/pull/2291)
+* [EXPORTER] Prometheus exporter emit resource attributes
+  [#2301](https://github.com/open-telemetry/opentelemetry-cpp/pull/2301)
 * [EXPORTER] Remove explicit timestamps from metric points exported by Prometheus
   [#2324](https://github.com/open-telemetry/opentelemetry-cpp/pull/2324)
 * [EXPORTER] Handle attribute key collisions caused by sanitation
@@ -27,6 +29,9 @@ Increment the:
   [#2224](https://github.com/open-telemetry/opentelemetry-cpp/pull/2224)
 * [REMOVAL] Drop C++11 support
   [#2342](https://github.com/open-telemetry/opentelemetry-cpp/pull/2342)
+* [EXPORTER] Add otel_scope_name and otel_scope_version labels to the prometheus
+  exporter.
+  [#2293](https://github.com/open-telemetry/opentelemetry-cpp/pull/2293)
 
 Important changes: